qdrant_indexer/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.10
qdrant-client==1.12.1
sentence-transformers==3.2.1
torch==2.4.1
